restructured in prep for deno
